Answer for Q 2 and 3
The ogg and mp3 and mp4 listed on your code WILL NOT work because you are using a relative path to a destination where the files are missing.

Remember that relative paths require you to have the resources requested inside your project folder. Add an mp3 and mp4 into your project and update the paths to reflect a resource that exists in your project.

The al olsen URL destination that you have listed doesn't exist... Try something like : http://alolsen.net/CUNY/citytech/may2016/

There is nothing to see in the directory named "citytech" which your code points to.