Q 4
You don't seem to have included a <a href="#top"> GO TO TOP</a> link anywhere in your code. You have, however, added that anchor destination by adding id="top" to the <head> tag. I would suggest adding this id to the <body> instead.

Also, it is very important to add your closing tags!!! You forgot to add the </body> and </html> closing tags at the end of this document.